refactor: split QueryCollectionService into specialized services
- Created LocalCollectionService for IndexedDB operations
- Created RemoteCollectionService for API operations
- Created CollectionUtilsService for import/export and tree utilities
- Main QueryCollectionService now delegates to specialized services
- Improved separation of concerns and code maintainability
- All 32 tests passing
- Reduced main service from 530+ lines to ~300 lines
