+++ packages/altair-app/src/app/modules/altair/altair.module.ts
@@ -87,6 +87,9 @@ const providers = [
   services.KeybinderService,
   services.StorageService,
   services.QueryCollectionService,
+  services.LocalCollectionService,
+  services.RemoteCollectionService,
+  services.CollectionUtilsService,
   services.EnvironmentService,
   services.PluginRegistryService,
   services.PluginEventService,
