+++ packages/altair-app/src/app/modules/altair/services/index.ts
@@ -8,6 +8,9 @@ export { ElectronAppService } from './electron-app/electron-app.service';
 export { KeybinderService } from './keybinder/keybinder.service';
 export { StorageService } from './storage/storage.service';
 export { QueryCollectionService } from './query-collection/query-collection.service';
+export { LocalCollectionService } from './query-collection/local-collection.service';
+export { RemoteCollectionService } from './query-collection/remote-collection.service';
+export { CollectionUtilsService } from './query-collection/collection-utils.service';
 export { EnvironmentService } from './environment/environment.service';
 export { PluginRegistryService } from './plugin/plugin-registry.service';
 export { PluginEventService } from './plugin/plugin-event.service';
